Peptide-level imputation in mass spectrometry label-free proteomics by low-rank approximation

Returns a completed peptide intensity matrix where missing values (NAs) are imputated by low-rank approximation of the input matrix. Non-NA entries remain unmodified. msImpute requires at least 4 non-missing measurements per peptide across all samples. It is assumed that peptide intensities (DDA), or MS1/MS2 normalised peak areas (DIA), are log2-transformed and normalised (e.g. quantile normalisation).

msImpute(
  object,
  rank.max = NULL,
  lambda = NULL,
  thresh = 1e-05,
  maxit = 100,
  trace.it = FALSE,
  warm.start = NULL,
  final.svd = TRUE
)

Arguments

object
Numeric matrix giving log-intensity where missing values are denoted by NA. Rows are peptides, columns are samples.

rank.max
Numeric. This restricts the rank of the solution. is set to min(dim(object)-1) by default.

lambda
Numeric. Nuclear-norm regularization parameter. Controls the low-rank property of the solution to the matrix completion problem. By default, it is determined at the scaling step. If set to zero the algorithm reverts to "hardImputation", where the convergence will be slower.

thresh
Numeric. Convergence threshold. Set to 1e-05, by default.

maxit
Numeric. Maximum number of iterations of the algorithm before the algorithm is converged. 100 by default.

trace.it
Logical. Prints traces of progress of the algorithm.

warm.start
List. A SVD object can be used to initialize the algorithm instead of random initialization.

final.svd
Logical. Shall final SVD object be saved? The solutions to the matrix completion problems are computed from U, D and V components of final SVD.

Value

Missing values are imputed by low-rank approximation of the input matrix. If input is a numeric matrix, a numeric matrix of identical dimensions is returned.

Details

msImpute operates on the softImpute-ALS algorithm. For more details on the underlying algorithm, please see softImpute package.

References

Hastie, T., Mazumder, R., Lee, J. D., & Zadeh, R. (2015). Matrix completion and low-rank SVD via fast alternating least squares. The Journal of Machine Learning Research, 16(1), 3367-3402.

Hediyeh-zadeh, S., Webb, A. I., & Davis, M. J. (2020). MSImpute: Imputation of label-free mass spectrometry peptides by low-rank approximation. bioRxiv.

See also

Examples

set.seed(101)
n=12000
p=10
J=5
np=n*p
missfrac=0.3
x=matrix(rnorm(n*J,mean = 5,sd = 0.2),n,J)%*%matrix(rnorm(J*p, mean = 5,sd = 0.2),J,p)+
  matrix(rnorm(np,mean = 5,sd = 0.2),n,p)/5
ix=seq(np)
imiss=sample(ix,np*missfrac,replace=FALSE)
xna=x
xna[imiss]=NA
keep <- (rowSums(!is.na(xna)) >= 4)
xna <- xna[keep,]
xna <- scaleData(xna)

#> bi-scaling ...
#> data scaled 

xcomplete <- msImpute(object=xna)

#> maximum rank is 9 
#> computing lambda0 ... 
#> lambda0 is 134.2294 
#> fit the low-rank model ... 
#> model fitted. 
#> Imputting missing entries ... 
#> Imputation completed